refactor: 예시를 oauthcontroller에 적용
