+++ src/main/java/com/kestrel/weddingbookkeeper/api/auth/vo/JwtToken.java
@@ -1,8 +1,11 @@
 package com.kestrel.weddingbookkeeper.api.auth.vo;
 
+import io.swagger.v3.oas.annotations.media.Schema;
+
+@Schema(description = "jwt 토큰")
 public class JwtToken {
 
-    private long id;
+    @Schema(description = "jwt 토큰 값")
     private String jwtToken;
 
     public JwtToken(String jwtToken) {
